fix(aot): resolve macOS ARM ABI tag mismatch and layout bugs
- fix(benchmarks): add 300ms timeout for QBE execution to avoid hangs
- ci: include bun in performance benchmarks across all OS targets
- fix(proton): explicitly set arm64_apple target on macOS ARM
- fix(aot): change ValueType to uint32_t to match QBE 'w' type and fix struct padding
- fix(aot): align QBE runtime and codegen tags with actual ValueType enum values
- fix(aot): update C++ runtime signatures to expect uint32_t for tags, fixing ARM64 variadic calling convention mismatches

+++ include/types/value.h
@@ -64,7 +64,7 @@ class Buffer;
  * This enum represents every type in the language. The JIT compiler
  * uses these for type specialization guards.
  */
-enum class ValueType : uint64_t {
+enum class ValueType : uint32_t {
     NIL,         ///< The absence of value (like null, but more existential)
     BOOLEAN,     ///< true or false - the building blocks of decision-making
     NUMBER,      ///< Double-precision floating point (because integers are overrated)
